+++ provider/datagen/src/registry.rs
@@ -253,7 +253,7 @@ registry!(
         LongYearRelativeTimeFormatDataV1Marker,
         ShortYearRelativeTimeFormatDataV1Marker,
         NarrowYearRelativeTimeFormatDataV1Marker,
-        //LongCompactDecimalFormatDataV1Marker,
+        LongCompactDecimalFormatDataV1Marker,
         ShortCompactDecimalFormatDataV1Marker,
     }
 );
